Review of your PR with some changes.
I'd also like to see some code snippets of both of the DNS SRV lookups i.e. from Lagom and Liberty. We need to explain the magic instead of the reader having to trawl through the code themselves. Hopefully, my Lagom/Akka SBR PR serves as inspiration in that regard.